+++ R/view.R
@@ -4,7 +4,7 @@
 #' `str_view()` shows the first match; `str_view_all()` shows all matches.
 #' Matches are surrounded by `<>` and coloured blue; unusual whitespace
 #' characters (i.e. everything apart from `" "` and `"\n"`) are shown with a
-#' red background.
+#' blue background.
 #'
 #' To build regular expressions interactively, check out the
 #' [RegExplain RStudio addin](https://www.garrickadenbuie.com/project/regexplain/).
@@ -12,10 +12,10 @@
 #' @param match If `TRUE`, shows only strings that match the pattern.
 #'   If `FALSE`, shows only the strings that don't match the pattern.
 #'   Otherwise (the default, `NA`) displays both matches and non-matches.
-#' @param html Use HTML output? If `TRUE` will create an HTML widget;
-#'   if `FALSE` will style using ANSI escapes. The default will prefers
-#'   ANSI escapes if available in the current terminal; you can override by
-#'   setting `option(stringr.html = TRUE)`.
+#' @param html Use HTML output? If `TRUE` will create an HTML widget; if `FALSE`
+#'   will style using ANSI escapes. The default prefers ANSI escapes if
+#'   available in the current terminal; you can override by setting
+#'   `options(stringr.html = TRUE)`.
 #' @param use_escapes If `TRUE`, all non-ASCII characters will be rendered
 #'   with unicode escapes. This is useful to see exactly what underlying
 #'   values are stored in the string.
@@ -43,9 +43,16 @@
 #' # Show all matches
 #' str_view_all(c("abc", "def", "fgh"), "d|e")
 str_view <- function(string, pattern = NULL, match = NA, html = NULL, use_escapes = FALSE) {
+  rec <- vctrs::vec_recycle_common(string = string, pattern = pattern)
+  string <- rec$string
+  pattern <- rec$pattern
+
   html <- str_view_use_html(html)
 
-  out <- str_view_filter(string, pattern, match)
+  filter <- str_view_filter(string, pattern, match)
+  out <- string[filter]
+  pattern <- pattern[filter]
+
   if (!is.null(pattern)) {
     out <- str_replace(out, pattern, str_view_highlighter(html))
   }
@@ -56,15 +63,22 @@ str_view <- function(string, pattern = NULL, match = NA, html = NULL, use_escape
     out <- str_view_special(out, html = html)
   }
 
-  str_view_print(out, html)
+  str_view_print(out, filter, html = html)
 }
 
 #' @rdname str_view
 #' @export
 str_view_all <- function(string, pattern = NULL, match = NA, html = NULL, use_escapes = FALSE) {
+  rec <- vctrs::vec_recycle_common(string = string, pattern = pattern)
+  string <- rec$string
+  pattern <- rec$pattern
+
   html <- str_view_use_html(html)
 
-  out <- str_view_filter(string, pattern, match)
+  filter <- str_view_filter(string, pattern, match)
+  out <- string[filter]
+  pattern <- pattern[filter]
+
   if (!is.null(pattern)) {
     out <- str_replace_all(out, pattern, str_view_highlighter(html))
   }
@@ -75,19 +89,19 @@ str_view_all <- function(string, pattern = NULL, match = NA, html = NULL, use_es
     out <- str_replace_all(out, fixed("\\u001b"), "\u001b")
   }
 
-  str_view_print(out, html)
+  str_view_print(out, filter, html = html)
 }
 
 str_view_filter <- function(x, pattern, match) {
   if (is.null(pattern)) {
-    x
+    rep(TRUE, length(x))
   } else {
     if (identical(match, TRUE)) {
-      x[str_detect(x, pattern)]
+      str_detect(x, pattern)
     } else if (identical(match, FALSE)) {
-      x[!str_detect(x, pattern)]
+      !str_detect(x, pattern)
     } else {
-      x
+      rep(TRUE, length(x))
     }
   }
 }
@@ -111,18 +125,18 @@ str_view_special <- function(x, html = TRUE) {
   if (html) {
     replace <- function(x) paste0("<span class='special'>", x, "</span>")
   } else {
-    replace <- function(x) cli::col_white(cli::bg_red(x))
+    replace <- function(x) cli::bg_cyan(x)
   }
 
-  # Highlight, anything whitespace characters that aren't a space.
+  # Highlight any non-standard whitespace characters
   str_replace_all(x, "[\\p{Whitespace}-- \n]+", replace)
 }
 
-str_view_print <- function(x, html = TRUE) {
+str_view_print <- function(x, filter, html = TRUE) {
   if (html) {
     str_view_widget(x)
   } else {
-    structure(x, class = "stringr_view")
+    structure(x, id = which(filter), class = "stringr_view")
   }
 }
 
@@ -152,7 +166,25 @@ str_view_widget <- function(lines) {
 }
 
 #' @export
-print.stringr_view <- function(x, ...) {
+print.stringr_view <- function(x, ..., n = 20) {
+  n_extra <- length(x) - n
+  if (n_extra > 0) {
+    x <- x[seq_len(n)]
+  }
+
+  id <- format(paste0("[", attr(x, "id"), "] "), justify = "right")
+  x <- paste0(id, x)
+  x <- str_replace_all(x, "\n", paste0("\n", strrep(" ", nchar(id[[1]]))))
+
   cat(x, sep = "\n")
+  if (n_extra > 0) {
+    cat("... and ", n_extra, " more\n", sep = "")
+  }
+
   invisible(x)
+}
+
+#' @export
+`[.stringr_view` <- function(x, i, ...) {
+  structure(NextMethod(), id = attr(x, "id")[i], class = "stringr_view")
 }